allow disabling eval patching
