+++ src/sdk/endpoint.ts
@@ -9,28 +9,6 @@ const SHORTHANDS: Record<string, string> = {
   global: "https://global.frcapi.com/api/v2/captcha",
 };
 
-export function getSDKAPIEndpoint(): string | undefined {
-  // 1. We check for the meta tag `frc-api-endpoint`
-  const m: HTMLMetaElement | null = document.querySelector(`meta[name="frc-api-endpoint"]`);
-  if (m) return m.content;
-
-  // 2. We check the current script element for `data-frc-api-endpoint`.
-  const cs = document.currentScript;
-  if (cs) {
-    const endpoint = cs.dataset["frcApiEndpoint"];
-    if (endpoint) return endpoint;
-  }
-
-  // 3. We search for widgets that specify `data-api-endpoint`.
-  const we = document.querySelector(".frc-captcha[data-api-endpoint]") as HTMLElement;
-  if (we) {
-    const endpoint = we.dataset["apiEndpoint"];
-    if (endpoint) return endpoint;
-  }
-
-  return undefined;
-}
-
 export function resolveAPIEndpoint(optionValue: string | undefined) {
   if (!optionValue) {
     // We default to the global endpoint
